*** Alteration Refactoring (NEW   IRegionManager  !!!)
Ran into a design issue using shared controls with nested regions. Shared region names became a problem. After consideration I decided to write a custom IRegionManager; and did so with a slightly new event aggregator design (small change to the event type). The POC is working; and I needed to check in all of these changes. The Alteration Effect UI is nearly completed; but this IRegionManager issue had to be dealt with first.