Brush copy function
Brushes are now read from the configuration while it's being mapped; and stored in the IScenarioCollectionProvider as a separate observable collection. There's a few new methods to provide changes to the brush collection; but it is ONLY CALCULATED - NOT STORED. So, now a user can just copy a brush to their animation from any of the ones stored in the scenario.