Worked on UI performance and rendering
The new layered level canvas allows separating content and layout - and seems to help rendering performance. Probably, putting all the light radii and auras on a single layer together - and then applying the opacity mask did the trick. We should be able to open up Light Radii (already done) for characters; and increase the size of the level some.