***Altered layout to use single cell walls (PREP FOR TERRAIN)
This one was a lot of work to prepare for putting terrain layer in the game. Most of it was to use single cell walls as opposed to edge-of-cell walls. The apprearance is temporary to show cells with some color; but Terrain SVG's will be added. The framework has been added - with an SVG folder and a couple of simple toy examples to test performance. Must more to go...