Changed linear corridor layout
Linear corridors are now laid out according to a simple "connecting rectangle" rule. The MST is calculated based on the "Roguian" distance between the region centers. Then, each edge creates a "rectangle" using the opposing centers as opposing corners. Then, the path from one to the other is created by tracing the rectangle.