Cleaned up math folder - integrated region graph calculation with the region class
Got rid of most of the routines that weren't being used. The region class now inherits from IRegionGraphWeightProvider - which is now specifically used for regions. The nearest neighbor calculation is now done right there - which the results stored for lookup (using the interface).

So, it's all pretty much tied together now for using regions to do direct calculations for layout generation  - which should be really useful.