Grouped regions into layers
Started generating terrain layers and realized the region maps were just layers - so grouped them by layer. So, there's one for rooms, and then a collection of layers for terrain - each with regions.

The LayerMap object has an index to find the associated regions. It also has a layer name to identify it with the scenario configuration resources - so that the layer cells can be rendered.