Delaunay Triangulation (re-worked!!!) :) :) :)
Completely componentized triangulation. This will allow creating corridors between the MST and the full Delaunay Triangulation.

The algorithm has been reworked to think of the graph as a graph over all region connections - treating each individual connection vertex as a graph node.

The result is the closest thing to a "Delaunay triangulation" of the actual regions that you'll get. And, this allows specifying the amount of "extra" corridors - between the MST and Delaunay - so that the graph is still minimally connected; but with extra connections between adjacent rooms.