Level Refactoring - Added Layout Queries!!!
Changed level to contain the LayoutGrid (was LevelGrid); and now a new content container LevelContent. The content container maintains the 2D content array for querying. The new layout grid has occupied queries added - per layer - and now contains 10 static layers along with the additional dynamic terrain layer collection.

Removed the reference dependency for the Graph<>; and replaced it with a string ID for the regions.

The Graph<> is now a Graph and is serialized with the CONNECTED layer(s). There is only ONE connected now because the triangulation routine is NON-DETERMINISTIC. This was previously a bug. So, you can't calculate the triangulation AFTER placing corridors. This is left strictly to the connection builder - which now returns the Graph for storage.

So, the "Connection" layer is now responsible for maintaining relationships between its regions. All other path finding is left to a Dijkstra scan.

Basically, there's no real way to easily know how regions are related to each other without running a Dijkstra scan. So, if we needed this data stored - it could be post-calcualted (after triangulation) using a Dijkstra scan.

Finally, the Layout Grid queries have added support for checking occupancy; AND for finding "Distant" locations based on a weighted random draw. These are now used to specify, further, how to place content. Example: Stairs up, Stairs down, and Save point now "prefer" to be farther apart. This is based on a simple PDF calculated from the distances of layer occupancy locations; and used for a single weighted random draw.