REFACTORED LOCATION STORAGE!!
Location propery removed from ScenarioObject (!!!)  This was to prevent duplicating data since there is a new need for occupancy data at the layout grid level.

The location data is now stored on the level content; and queried via the Level.

This means that the Location "property" is no longer "loose"; and is more tightly managed. So, any other components that store the "Location" must be kept closer to the Level / Layout Grid. So, probably, the occupation data will be moved - one more time - to a separate layer data structure inside the layout grid.