Moved "Explored" content to "Memorized"
New feature added to handle ScenarioObject.IsExplored flag.. mostly because it doesn't really make sense to have "Explored" an object that can be removed by an enemy character.

So, the LevelContent object is now call ContentGrid and has two instances inside the Level: one for normal content; and the other for memorized content.

Added minimal design changes to keep these separated clearly and to display the memorized content using what was the "explored" visual layer (which was the same layer as all the objects.. just different rendering)