Connection layer refactoring
Moved the connection data for regions into a sub-class with a corresponding layer sub-class. This separated the connection information for regions entirely; and forced a refactoring of the layout generation code to specify whether connected or non-connected regions were being created.

The connected regions / layer are going to house more pre-calculated data. So, this refactoring was meant to prepare for minimizing additional data storage; and cleaning up the Region<> code.