Working Bitmap rendering using WriteableBitmap !!!!!
Performance issue solved!!!!  This method of rendering is going to replace the entire layout rendering process.

A 3rd cache has been created to store rendered bitmaps at the specified zoom scale. This is called from the LevelCanvas to render the DrawingImage instances - which are also cached - on a different layer.