Applied Vision Intensity Falloff
The rendering looks beautiful!!!  After working with WriteableBitmap instances to fix the performance bottleneck - the baseline (that's on right now) is pretty acceptable - though a little laggy. It's rendering the entire layout each step!

Using the vision falloff blew up the cache size; but that's easily fixed by discretizing the lighting for each cell.