Setup lighting "pipeline" OPTIMIZED CACHE HASHING
Pipeline applies light effect in order of storage with the GridCell.

New cache hashing WAS MAJOR PERFORMANCE FIX! For whatever reason, the dictionary hashsing of the cach key struct wasn't unique to its properties (!!!) and it was SLOW.

So, just created a hash manually from the input and prevented any memory allocation. The cache size went from 1500 -> 80 (per level-ish) (!!!!!!!)  So, that was a HUGE memory leak; and performance was terrible..

MUCH BETTER NOW