FINISHED LIGHTING!!!! :) :) :) :)
FINALLY FOUND A LIGHTING ALGORITHM THAT WORKS!!!

Used weighted averages of the color components - while normalizing the lightness. Effectively, it made a color filter - which then needed to be be darkened according to the effective light intensity at each cell.

This preserved the original color channel "bias" (for the most part) - which was difficult. Using HSL coordinates was the way to go in the end.