***SERIALIZATION PROBLEM*** Refactored Layout Generation - TRIANGULATION FIXED
SERIALIZATION PROBLEM:  Had to be checked in with a MS BUG!!! Region<T> serialization ctor NOT BEING CALLED! NO EXPLANATION! So, the game isn't playable; but the editor is functioning.

NEW SERIALIZER IN THE WORKS!

Scaling of the triangulator wasn't using the rendered distance! So, that was fixed; and the connection layer was updated with more careful planning of the rooms. ConnectionRoom enum was added to serve as the base region layer. This is setup in the connection layer; but the regions don't have to include the corridors. (they don't).

So, the layout generator was re-vamped with more control given to the layout container; and the graph algorithms are overhauled with lots of planning given to the connection graph with careful consideration to route planning for the connection layer.